requirements.txt
pyjwt==1.5.0
cryptography==1.8.1